requirement.txt
Flask>=2.0
flask-socketio>=5.0
python-dotenv
eventlet
flask-sqlalchemy